order-service/requirements.txt
Flask==2.3.2
requests
psycopg2-binary